[UNTESTED] Pretty sure that's all the floats removed from variadics

+++ zephyrapp/game/src/monitors/cat_monitor_sparklines.c
@@ -66,11 +66,11 @@ void CAT_monitor_render_sparklines()
 
 	cursor_y = SPARKLINE_MAX_Y + 8;
 	CAT_set_text_colour(CAT_WHITE);
-	cursor_y = CAT_draw_textf(SPARKLINE_X, cursor_y, "Max score: %s \1 %.0f/100\n", CAT_AQ_get_grade_string(max_score), max_score*100)+8;
+	cursor_y = CAT_draw_textf(SPARKLINE_X, cursor_y, "Max score: %s \1 %d/100\n", CAT_AQ_get_grade_string(max_score), (int)(max_score*100))+8;
 	CAT_set_text_colour(CAT_WHITE);
-	cursor_y = CAT_draw_textf(SPARKLINE_X, cursor_y, "  Mean score: %s \1 %.0f/100\n", CAT_AQ_get_grade_string(mean_score), mean_score*100)+8;
+	cursor_y = CAT_draw_textf(SPARKLINE_X, cursor_y, "  Mean score: %s \1 %d/100\n", CAT_AQ_get_grade_string(mean_score), (int)(mean_score*100))+8;
 	CAT_set_text_colour(CAT_WHITE);
-	cursor_y = CAT_draw_textf(SPARKLINE_X, cursor_y, "    Min score: %s \1 %.0f/100\n", CAT_AQ_get_grade_string(min_score), min_score*100)+8;
+	cursor_y = CAT_draw_textf(SPARKLINE_X, cursor_y, "    Min score: %s \1 %d/100\n", CAT_AQ_get_grade_string(min_score), (int)(min_score*100))+8;
 }
 
 void CAT_monitor_MS_sparklines(CAT_machine_signal signal)
